pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpha-pipeline"
version = "0.1.0"
description = "Prediction market alpha pipeline — orderbook ingestion and feature computation"
requires-python = ">=3.11"
dependencies = [
    "websockets>=13.0",
    "python-socketio[asyncio]>=5.11",
    "aiohttp>=3.9",
    "orjson>=3.9",
    "polars>=1.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "uvloop>=0.19; sys_platform != 'win32'",
    "structlog>=24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
advanced = ["pynufft>=2024.0", "py-vollib-vectorized"]
analysis = ["matplotlib>=3.8"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha_pipeline"]
